requirements.txt
Django>=3.0,<4.0
psycopg2-binary>=2.8
StringGenerator==0.4.4
# django-plotly-dash==1.6.3
djangorestframework
markdown
django-filter
requests
PyYAML
httpx